app/requirements.txt
azure-storage-blob
Pillow
st7735
bme280
pms5003
enviroplus
ltr559
fonts